Development

+++ src/pages/Settings.jsx
@@ -4,6 +4,7 @@ import { useState } from "react";
 import Alert from "../components/Alert";
 import Card from "../components/Card";
 import FancyRadioButtonGroup from "../components/Form/FancyRadioButtonGroup";
+import Input from "../components/Form/Input";
 import HR from "../components/HR";
 import Modal from "../components/Modal";
 import Page from "../components/Page";
@@ -28,6 +29,9 @@ export default function Settings({ deferredPrompt }) {
     const [showImportModal, setShowImportModal] = useState(false)
     const [showDeleteModal, setShowDeleteModal] = useState(false)
     const [showTestDataModal, setShowTestDataModal] = useState(false)
+    const [confirmInput, setConfirmInput] = useState('')
+    const confirmText = 'test' /* 'I will lose my current data' */
+    const hasCorrectConfirmText = confirmInput.toLowerCase() === confirmText.toLowerCase() || !dbHasData
 
     const [lastExportDate, setLastExportDate] = useLocalStorage('lastExportDate', getCurrentDateTime(true))
     const promoteExport = (new Date(getCurrentDateTime(true)) - new Date(lastExportDate)) / (1000 * 60 * 60 * 24) >= 7
@@ -52,27 +56,37 @@ export default function Settings({ deferredPrompt }) {
     }
 
     const handleImport = () => {
+        if (!hasCorrectConfirmText) return
         promptFileSelection((file) => {
             importJsonFileToDatabase(file, () => {
                 setDbData(db.getAllData())
                 setLastExportDate(getCurrentDateTime(true))
-                setShowImportModal(false)
+                closeModals()
             })
         }, { accept: '.json' })
     }
 
     const handleDeleteAllData = () => {
+        if (!hasCorrectConfirmText) return
         db.deleteAllData()
         setLastExportDate(getCurrentDateTime(true))
         setDbData(db.getAllData())
-        setShowDeleteModal(false)
+        closeModals()
     }
 
     const handleSetTestData = () => {
+        if (!hasCorrectConfirmText) return
         db.setTestData()
         setLastExportDate(getCurrentDateTime(true))
         setDbData(db.getAllData())
+        closeModals()
+    }
+
+    const closeModals = () => {
+        setShowImportModal(false)
+        setShowDeleteModal(false)
         setShowTestDataModal(false)
+        setConfirmInput('')
     }
 
     return (
@@ -83,7 +97,6 @@ export default function Settings({ deferredPrompt }) {
                 </Card>
 
                 {!isPWA && (
-                    // TODO: dont show if deferredPrompt is null. Then animate in when it available
                     <Card title="Download">
                         <p className="mb-2">Download the app to use it offline</p>
                         <button className={`${deferredPrompt ? 'btn-primary' : 'btn-secondary'} w-full`} disabled={!deferredPrompt} onClick={handleInstall}>Install{!deferredPrompt && <Spinner className="ml-2" />}</button>
@@ -97,9 +110,7 @@ export default function Settings({ deferredPrompt }) {
                         <button className={`${promoteExport ? 'btn-primary' : 'btn-secondary'} w-full`} onClick={handleExport}>Export Data<FontAwesomeIcon icon={faUpload} className="ml-2" /></button>
                         <button className="btn-secondary w-full" onClick={() => setShowImportModal(true)}>Import Data<FontAwesomeIcon icon={faDownload} className="ml-2" /></button>
                     </div>
-                    {dbHasData && (
-                        <p className="mt-2">Your last export was <strong>{timeDifferenceToText(lastExportDate)}</strong></p>
-                    )}
+                    {dbHasData && <p className="mt-2">Your last export was <strong>{timeDifferenceToText(lastExportDate)}</strong></p>}
                     <Alert message={promoteExport ? 'It is adviced to export your data every 7 days' : null} isCloseable={false} className="mt-2" />
 
                     <HR className="-my-1" />
@@ -109,40 +120,46 @@ export default function Settings({ deferredPrompt }) {
                     <button className="btn-danger w-full" onClick={() => setShowTestDataModal(true)}>Replace data with testdata</button>
                 </Card>
 
-                <Modal showModal={showImportModal} onClose={() => setShowImportModal(false)} title="Import data">
+                <Modal showModal={showImportModal} onClose={closeModals} title="Import data">
                     <p>Are you sure you want to import data? This will overwrite all your current data.</p>
-                    {dbHasData && (
+                    {dbHasData && <>
                         <p>Your last export was <strong>{timeDifferenceToText(lastExportDate)}</strong></p>
-                    )}
-                    <p className="font-semibold text-danger">This action cannot be undone</p>
+                        <p>Type &quot;<strong>{confirmText}</strong>&quot; to confirm</p>
+                        <Input label="Confirm" type="text" value={confirmInput} onChange={(e) => setConfirmInput(e.target.value)} className="mb-2" />
+                    </>}
                     <div className="flex gap-4">
-                        <button className="btn-secondary w-full" onClick={() => setShowImportModal(false)}>Cancel</button>
-                        <button className="btn-danger w-full" onClick={handleImport}>Import data</button>
+                        <button className="btn-secondary w-full" onClick={closeModals}>Cancel</button>
+                        <button className="btn-danger w-full" disabled={!hasCorrectConfirmText} onClick={handleImport}>Import data</button>
                     </div>
-                </Modal >
+                    <p className="font-semibold text-danger text-right">This action cannot be undone</p>
+                </Modal>
 
-                <Modal showModal={showDeleteModal} onClose={() => setShowDeleteModal(false)} title="Delete all data">
+                <Modal showModal={showDeleteModal} onClose={closeModals} title="Delete all data">
                     <p>Are you sure you want to delete all data?</p>
-                    {dbHasData && (
+                    {dbHasData && <>
                         <p>Your last export was <strong>{timeDifferenceToText(lastExportDate)}</strong></p>
-                    )}
-                    <p className="font-semibold text-danger">This action cannot be undone</p>
+                        <p>Type &quot;<strong>{confirmText}</strong>&quot; to confirm</p>
+                        <Input label="Confirm" type="text" value={confirmInput} onChange={(e) => setConfirmInput(e.target.value)} className="mb-2" />
+                    </>}
                     <div className="flex gap-4">
-                        <button className="btn-secondary w-full" onClick={() => setShowDeleteModal(false)}>Cancel</button>
-                        <button className="btn-danger w-full" onClick={handleDeleteAllData}>Delete data</button>
+                        <button className="btn-secondary w-full" onClick={closeModals}>Cancel</button>
+                        <button className="btn-danger w-full" disabled={!hasCorrectConfirmText} onClick={handleDeleteAllData}>Delete data</button>
                     </div>
-                </Modal >
+                    <p className="font-semibold text-danger text-right">This action cannot be undone</p>
+                </Modal>
 
-                <Modal showModal={showTestDataModal} onClose={() => setShowTestDataModal(false)} title="Replace data with test data">
+                <Modal showModal={showTestDataModal} onClose={closeModals} title="Replace data with test data">
                     <p>Are you sure you want to replace all data with test data? This will overwrite all your current data.</p>
-                    {dbHasData && (
+                    {dbHasData && <>
                         <p>Your last export was <strong>{timeDifferenceToText(lastExportDate)}</strong></p>
-                    )}
-                    <p className="font-semibold text-danger">This action cannot be undone</p>
+                        <p>Type &quot;<strong>{confirmText}</strong>&quot; to confirm</p>
+                        <Input label="Confirm" type="text" value={confirmInput} onChange={(e) => setConfirmInput(e.target.value)} className="mb-2" />
+                    </>}
                     <div className="flex gap-4">
-                        <button className="btn-secondary w-full" onClick={() => setShowTestDataModal(false)}>Cancel</button>
-                        <button className="btn-danger w-full" onClick={handleSetTestData}>Set test data</button>
+                        <button className="btn-secondary w-full" onClick={closeModals}>Cancel</button>
+                        <button className="btn-danger w-full" disabled={!hasCorrectConfirmText} onClick={handleSetTestData}>Set test data</button>
                     </div>
+                    <p className="font-semibold text-danger text-right">This action cannot be undone</p>
                 </Modal>
             </Page>
         </>
